checkpoint on association interface (DO NOT USE)

+++ viper_orchestrator/visintent/tracking/db_utils.py
@@ -10,20 +10,24 @@ from sqlalchemy.orm import Session, DeclarativeBase
 
 from viper_orchestrator.db import OSession
 from viper_orchestrator.db.table_utils import image_request_capturesets
-from viper_orchestrator.visintent.tracking.forms import RequestForm
+from viper_orchestrator.visintent.tracking.forms import (
+    RequestForm,
+    JunctionForm,
+)
 
 
-def _construct_associations(association_data, form, row):
+def _construct_associations(form: JunctionForm, row: DeclarativeBase):
     associations, removed = [], []
     with OSession() as session:
-        for field, rules in form.associated_tables.items():
-            setattr(form, rules['pivot'][1], getattr(row, rules['pivot'][1]))
+        for table, rules in form.association_rules.items():
+            setattr(form, rules["pivot"][1], getattr(row, rules["pivot"][1]))
             existing = form.get_associations(rules, session)
-            relations = {i: r for i, r in enumerate(association_data[field])}
+            relations = {i: r for i, r in enumerate(form.associated[table])}
             for junc_row in existing:
                 matches = [
-                    k for k, v in relations.items()
-                    if v == getattr(junc_row, rules['junc_pivot'][0])
+                    k
+                    for k, v in relations.items()
+                    if v == getattr(junc_row, rules["junc_pivot"])
                 ]
                 if len(matches) == 0:
                     removed.append(junc_row)
@@ -39,26 +43,13 @@ def _construct_associations(association_data, form, row):
                     associations.append(junc_row)
             # leftovers are new relations
             for v in relations.values():
-                junc_row = rules['junc']()
+                junc_row = table()
                 for attr, val in v.items():
                     setattr(junc_row, attr, val)
                 # this should already be set on existing junc table rows
-                setattr(junc_row, rules['self_attr'], row)
+                setattr(junc_row, rules["self_attr"], row)
                 associations.append(junc_row)
     return associations, removed
-
-
-def _separate_associations(form):
-    row_data, association_data = {}, {}
-    for k, v in form.cleaned_data.items():
-        if (
-            hasattr(form, "associated_tables")
-            and form.associated_tables.get(k) is not None
-        ):
-            association_data[k] = v
-        else:
-            row_data[k] = v
-    return row_data, association_data
 
 
 # TODO: excessively baroque
@@ -73,11 +64,9 @@ def _create_or_update_entry(
     helper function for processing data from bound Forms into DeclarativeBase
     objects
     """
-    data, association_data = _separate_associations(form)
+    data = {k: v for k, v in form.cleaned_data.items()}
     if extra_attrs is not None:
-        data |= {
-            attr: getattr(form, attr) for attr in extra_attrs
-        }
+        data |= {attr: getattr(form, attr) for attr in extra_attrs}
     try:
         # if this is an existing entry -- as determined by the
         # specified pivot field, which should have been extensively validated
@@ -120,13 +109,12 @@ def _create_or_update_entry(
             callobj = form.table_class
         row = callobj(**(keyfilter(lambda attr: attr in valid, data)))
     row.request_time = form.request_time
-    if len(association_data) > 0:
-        associations, removed = _construct_associations(
-            association_data, form, row
-        )
-    else:
-        associations, removed = [], []
-    session.add_all([row, *associations])
-    for r in removed:
-        session.delete(r)
-    return row, associations
+    if isinstance(form, JunctionForm):
+        for table, rows in form.associated.items():
+            if len(rows) == 0:
+                continue
+            associations, removed = _construct_associations(form, table, row)
+            session.add_all([row, *associations])
+            for r in removed:
+                session.delete(r)
+    return row
